arcindex/requirements.txt
# Runtime dependencies for the Arcindex CLI scaffolding.
click>=8.1
pyyaml>=6.0
